Move NewsblurUrlReader into a separate file
