+++ mk/mk.deps
@@ -98,12 +98,12 @@ src/controller.o: src/controller.cpp include/controller.h include/cache.h \
  include/inoreaderapi.h include/inoreaderurlreader.h \
  include/itemrenderer.h include/htmlrenderer.h include/textformatter.h \
  include/logger.h include/newsblurapi.h rss/feed.h rss/item.h \
- include/ocnewsapi.h include/oldreaderapi.h include/opmlurlreader.h \
- include/regexmanager.h include/remoteapi.h include/rssparser.h \
- include/stflpp.h include/strprintf.h include/ttrssapi.h \
- 3rd-party/json.hpp include/utils.h include/view.h include/controller.h \
- include/filebrowserformaction.h include/formaction.h include/history.h \
- include/keymap.h include/stflpp.h
+ include/newsblururlreader.h include/ocnewsapi.h include/oldreaderapi.h \
+ include/opmlurlreader.h include/regexmanager.h include/remoteapi.h \
+ include/rssparser.h include/stflpp.h include/strprintf.h \
+ include/ttrssapi.h 3rd-party/json.hpp include/utils.h include/view.h \
+ include/controller.h include/filebrowserformaction.h \
+ include/formaction.h include/history.h include/keymap.h include/stflpp.h
 src/dialogsformaction.o: src/dialogsformaction.cpp \
  include/dialogsformaction.h include/formaction.h include/history.h \
  include/keymap.h include/configparser.h include/configactionhandler.h \
@@ -304,14 +304,14 @@ src/matcherexception.o: src/matcherexception.cpp \
  include/matcherexception.h config.h include/strprintf.h
 src/newsblurapi.o: src/newsblurapi.cpp include/newsblurapi.h \
  include/remoteapi.h include/configcontainer.h include/configparser.h \
- include/configactionhandler.h rss/feed.h rss/item.h include/urlreader.h \
- include/remoteapi.h include/strprintf.h include/utils.h include/logger.h \
- config.h include/strprintf.h
-src/newsblururlreader.o: src/newsblururlreader.cpp include/newsblurapi.h \
- include/remoteapi.h include/configcontainer.h include/configparser.h \
- include/configactionhandler.h rss/feed.h rss/item.h include/urlreader.h \
+ include/configactionhandler.h rss/feed.h rss/item.h include/remoteapi.h \
+ include/strprintf.h include/utils.h include/logger.h config.h \
+ include/strprintf.h
+src/newsblururlreader.o: src/newsblururlreader.cpp \
+ include/newsblururlreader.h rss/feed.h rss/item.h include/urlreader.h \
  include/fileurlreader.h include/logger.h config.h include/strprintf.h \
- include/utils.h include/logger.h
+ include/remoteapi.h include/configcontainer.h include/configparser.h \
+ include/configactionhandler.h include/utils.h include/logger.h
 src/ocnewsapi.o: src/ocnewsapi.cpp include/ocnewsapi.h \
  include/remoteapi.h include/configcontainer.h include/configparser.h \
  include/configactionhandler.h rss/feed.h rss/item.h include/urlreader.h \
@@ -430,7 +430,7 @@ src/rssparser.o: src/rssparser.cpp include/rssparser.h \
  config.h include/strprintf.h rss/feed.h rss/item.h include/cache.h \
  include/configcontainer.h include/htmlrenderer.h include/textformatter.h \
  include/regexmanager.h include/logger.h include/newsblurapi.h \
- include/urlreader.h include/ocnewsapi.h include/rss.h rss/exception.h \
+ include/ocnewsapi.h include/urlreader.h include/rss.h rss/exception.h \
  rss/parser.h include/remoteapi.h rss/feed.h include/strprintf.h \
  rss/rssparser.h include/ttrssapi.h 3rd-party/json.hpp include/cache.h \
  include/utils.h
