+++ src/newsblururlreader.cpp
@@ -1,7 +1,8 @@
-#include "newsblurapi.h"
+#include "newsblururlreader.h"
 
 #include "fileurlreader.h"
 #include "logger.h"
+#include "remoteapi.h"
 #include "utils.h"
 
 namespace newsboat {
